refactor: extract shared addMonths and ACTIVE_STATUSES to dashboardUtils
